Break circular dependency main ↔ txdb
